adding household type as a variable to the ISurveyHousehold interface.
(This is not the best solution, but adapting the generic parameter in population synthesis would cause a rat tail of changes, so this option needs to be done in a time frame where more time is available)

+++ src/test/kotlin/domain/synthesis/householdgeneration/AbstractSynthesisTest.kt
@@ -190,13 +190,13 @@ open class SynthesisTest {
         }
 
         companion object {
-            var counter: Int = 0
+            var counter: Long = 0
                 get() = field++
                 private set
         }
     }
     protected class HouseholdBuilder<T> {
-        var id: Int = 0
+        var id: Long = 0
         var income = 0.euros
         val members: MutableList<SurveyPerson<out T>> = mutableListOf()
